requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=3.0
passlib>=1.7
python-dotenv>=1.0
marshmallow>=3.20
flask-marshmallow>=0.15
python-dateutil>=2.8